Extract trajectory math
